batch: Move binary file storage
Batch storage keeps text and atomic file persistence flows, and content
commit publication lives in its own module.

Binary batch persistence has different source-capture rules from text files
and submodule pointers. Housing it in storage keeps callers tied to a mixed
persistence module when they only need binary batch publication.

This commit begins separating atomic file implementations by moving binary
batch persistence into batch.binary_file_storage and updating production
callers.

Subsequent commits will move binary tests and boundary checks to the new
owner.
